Normalized Service and added request validation

+++ src/UrlShortener.Application/ShortUrlService.cs
@@ -10,9 +10,9 @@ namespace UrlShortener.Application;
 public interface IShortUrlService
 {
     Task<Either<RetrieveError, string>> GetById(string id);
-    Task<bool> DeleteById(string id);
+    Task<Either<DeleteError, Unit>> DeleteById(string id);
     Task<Either<CreateError, CreateResponse>> Create(string id, CreateRequest request);
-    Task<CreateResponse> Create(CreateRequest request);
+    Task<Either<CreateError, CreateResponse>> Create(CreateRequest request);
 }
 
 public class ShortUrlService(
@@ -36,24 +36,33 @@ public class ShortUrlService(
         return result;
     }
 
-    public async Task<bool> DeleteById(string id)
+    public async Task<Either<DeleteError, Unit>> DeleteById(string id)
     {
-        logger.LogInformation("DeleteById called");
-
         var entity = await repository.GetById(id);
+        if (entity == null)
+        {
+            logger.LogInformation("Entity not found, Id:{Id}", id);
+
+            return Either<DeleteError, Unit>.Left(DeleteError.NotFound);
+        }
         
         await repository.Delete(entity);
 
-        return true;
+        return Either<DeleteError, Unit>.Right(Unit.Default);
     }
 
     public async Task<Either<CreateError, CreateResponse>> Create(string id, CreateRequest request)
     {
-        logger.LogInformation("Create() called");
-
+        var validationError = ValidateCreate(request);
+        if (validationError != null)
+        {
+            return Either<CreateError, CreateResponse>.Left(validationError.Value);
+        }
+        
         var entity = await repository.GetById(id);
         if (entity != null)
         {
+            logger.LogInformation("Tried to override existing entity, Id:{Id}", id);
             var error = IsInvalid(entity)
                 ? CreateError.AlreadyExistsInvalid
                 : CreateError.AlreadyExists;
@@ -61,15 +70,30 @@ public class ShortUrlService(
             return Either<CreateError, CreateResponse>.Left(error);
         }
 
-        entity = new ShortUrl
+        var result = await CreateInternal(id, request);
+
+        return Either<CreateError, CreateResponse>.Right(result);
+    }
+
+    public async Task<Either<CreateError, CreateResponse>> Create(CreateRequest request)
+    {
+        var validationError = ValidateCreate(request);
+        if (validationError != null)
         {
-            Id = id,
-            CreatedAt = systemClock.UtcNow,
-            Url = request.Url,
-            TtlMinutes = request.TtlMinutes,
-        };
-        await repository.Create(entity);
-        var result = Map(entity);
+            return Either<CreateError, CreateResponse>.Left(validationError.Value);
+        }
+
+        var id = generator.Generate(request.Url!);
+        var entity = await repository.GetById(id);
+        if (entity != null)
+        {
+            // TODO: come up with a way of handling collisions
+            logger.LogInformation("Generated Id:{Id} that matches existing entity", id);
+            
+            return Either<CreateError, CreateResponse>.Left(CreateError.AlreadyExists);
+        }
+        
+        var result = await CreateInternal(id, request);
 
         return Either<CreateError, CreateResponse>.Right(result);
     }
@@ -77,13 +101,36 @@ public class ShortUrlService(
     private static CreateResponse Map(ShortUrl e) => 
         new (e.Id, e.Url, e.TtlMinutes, e.CreatedAt);
 
-    public Task<CreateResponse> Create(CreateRequest request)
+    private async Task<CreateResponse> CreateInternal(string id, CreateRequest request)
     {
-        logger.LogInformation("Create() called");
-
-        var id = generator.Generate(request.Url);
+        var entity = new ShortUrl
+        {
+            Id = id,
+            CreatedAt = systemClock.UtcNow,
+            Url = request.Url!,
+            TtlMinutes = request.TtlMinutes,
+        };
+        await repository.Create(entity);
+        var result = Map(entity);
         
-        return Task.FromResult(new CreateResponse(id, request.Url, request.TtlMinutes, systemClock.UtcNow));
+        return result;
+    }
+
+    private CreateError? ValidateCreate(CreateRequest request)
+    {
+        if (string.IsNullOrEmpty(request.Url))
+        {
+            logger.LogInformation("Url is missing");
+            return CreateError.MissingUrl;
+        }
+        
+        if (!Uri.TryCreate(request.Url, UriKind.Absolute, out _))
+        {
+            logger.LogInformation("Badly formatted url: {Url}", request.Url);
+            return CreateError.BadUrl;
+        }
+        
+        return null;
     }
 
     private bool IsInvalid(ShortUrl entity)
